Problem 500 and in progress 501
Yay, solved one of the most recent problems!
501 has a correct implementation that is too slow, and another
incomplete implementation that I'm working on.

+++ problemsupport.cpp
@@ -11,7 +11,7 @@ int numDigits(int x) {
 	return floor(log10(x)) + 1;
 }
 
-int numDivisors(int x) {
+int numDivisors(__int64 x) {
 	if (x < 0) return 0;
 	if (x == 1) return 1;
