+++ problemsupport.h
@@ -4,7 +4,7 @@
 int getDigit(int x, int pos);
 int getNextPrime(int x);
 int numDigits(int x);
-int numDivisors(int x);
+int numDivisors(__int64 x);
 bool isPalindrome(int x);
 bool isPentagonal(__int64 x);
 bool isPrime(__int64 x);
